+++ src/example/app/app.component.ts
@@ -1,4 +1,5 @@
-import { Component } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
+import { environment } from "../environments/environment";
 
 @Component({
   selector: "app-root",
@@ -7,4 +8,11 @@ import { Component } from "@angular/core";
 })
 export class AppComponent {
   title = "ngx-fileupload";
+
+  public disableAnimations = false;
+
+  constructor() {
+    console.log(environment);
+    this.disableAnimations = environment.disableAnimations || false;
+  }
 }